Lock - Don't keep blocking after shutdown
LockServiceImpl can close(), but even then, it can have
threads that block indefinitely, which will probably never
wake up, since nothing will ever unlock the locks they
depend on. To fix this, keep a set of all such threads and
interrupt them when shutting down.

This is probably most relevant when using clustered lock: a
re-election in the middle of a blocking call will cause the
thread to wait forever, which will cause the client to
hang forever, consume a thread permanently, and prevent the
lock service's memory from getting GC'ed.