Timestamp - don't ask for bogus number of timestamps
The timestamp request batching logic can get into a bad
state if it gets an exception talking to the timestamp
service. In that case, the number of timestamps to get has
been lost, overwritten by a negative number to also prevent
further threads from joining the batch. In this case, give
up on that batch.