Clean up executors
Also, fix a possible race condition, and mitigate a potential issue
whereby too many tasks are submitted to a ParallelTask, and instead
of doing the sane thing and blocking, it throws a RejectedExecutionException.

I'm unsure how this edge case happens though, since BlockingWorkerPool
is synchronized on the most important methods, so maybe the state
variable that was being tracked wasn't visible to all threads immediately.