Leader election - less contentious leader checks
Leader-checking calls were batched in a way that a lot of
synchronization. This change removes a lot of that. The
idea is that whichever thread creates a new batch is
responsible for actually making a remote call and writing
down the result, so now there's no need to synchronize to
figure out whose job that is. Now the only blocking is by
threads that want to see that result.