don't bother truncating tables that are already empty
This is to reduce schema mutations further, mostly because of
people still using Cassandra 1.2, which has a terribly overly coarse
lock that makes truncates while compactions are happening trivial to
wait forever and timeout.

There is a fringe legitimate use case that this infringes upon, which is
people aware of tombstoned or TTL-expired data in their tables below
the atlas transaction/timestamp layer and who want this garbage cleaned
up before the next compaction by calling truncate manually.