Fix a bug where postgres would attempt to insert too many values at once, causing it to error and close the connection.
This solution is probably not generalized -- postgres should be able to handle 32767 values at once, but in my testing on a single use case, it seemed to only be able to handle ~8000. I think this may have been because my atlas schema had 4 values. The constant I put in is a guess that unblocked my use case.