InMemoryKVS no longer violates ACID
At present, IMKVS does not copy arrays that are written to it. This means that
you are liable to see bad behaviour by doing expected safe things like
modifying an array that you used in a transaction (after the transaction
is committed?!)

InMemoryKVS. As much as the InMemoryKVS is designed for testing only,
it's still pretty rubbish. For example, it meant that #288 was exhibited
on the streamstore method that wrote in multiple transactions when using
an in-memory kvs.