safely allow moving between RF=1 and RF=3
The issue at heart is that a many node cluster
moving from RF=1 to RF=3 fully online will have
to move through RF=2, but temporarily won't have
the strong consistency guarantees normally afforded
by RF=2 'quorum' happening to be the same as 'all' in
a single-DC cluster. (a quorum of 2 is 2)

PDS-43637 is the person who would need this change.