Fix concurrency bug that allows writes to happen after putWithTimestamps returns
`PartitionedKeyValueService.putWithTimestamps` starts replicationFactor-many Futures, each calling `putWithTimestamps`
on the underlying KVS. However, the method only blocks on `completeWriteRequest`, which returns as soon as
writeFactor-many of the Futures succeed. Therefore it was possible that replicationFactor - writeFactor of the
Futures complete after putWithTimestamps already returned.

This commit adds syncrhronization to ensure that all started Futures complete before putWithTimestamps returns.

It is possible that some of the futures finish executing only after the function returns
because completeWriteRequest returns after writeFactor-many