QA-95214 Check for repeated insert of IDs
We sometimes see primary key violations on pt_temp_ids in testing.
This commit tests the theory that our SQL insert statement is somehow occurring twice,
perhaps due to an infrastructure issue.  This commit only takes effect when assertions are
enabled, so it should have no effect on production environments.