Don't create a separate cassandra pool to healthcheck cassandra hosts
Extra pools cause noise with jmx metrics (each pool has its own set of
metrics which aren't removed from the MBean server on shutdown).

This is also unnecessary since we should have a pool for each host.
Technically, the existing code did not maintain that invariant (which
implicitly resulted in it healthchecking blacklisted hosts that were no
longer in the config). Maintaining that invariant adds a small
complication (we _must_ remove a host from blacklistedHosts before
removing the corresponding pool from currentPools) but it doesn't seem
too bad.

There is also a check-then-act race where we might grab a blacklisted
host, remove it, and then try and check its health. Currently this is
not an issue because both actions are done inside a synchronized
refreshPools call, but it's sketchy. I think the real solution to this
is a r/w lock on currentPools because, as far as I can tell, we also
have this issue with general currentPools usage (e.g. runOnHost).

(cherry picked from commit 013fdb2337e610898d56b19ae325bda6acd219e2)