Improvements to ConflictDetectionManager.
ConflictDetectionManager previously didn't handle lots of tables well.
Every call to get() would involve creating a new copy of the
table->conflict handler map, and get() is called once for every
SerializableTransaction#get().

This refactors the class so that a new map copy is only created on config
modification, and calls to get() return a precomputed ImmutableMap.