removing console, adding null check
Removing the bit about atlasdb console because it doesn't include all the dependencies for CLIs, and so clients using the bundle could run into issues if they try using console now, and then want to use future clis. We plan on getting this in soon, so holding off on documenting this.