ssl: false should override the presence of an SslConfiguration
We want to add a default SslConfiguration to our internal application.
However, users may have explicitly disabled Ssl for communication with
their Cassandra, which will be broken by the presence of the new block
overriding the `ssl()` param..

Despite being deprecated, the `ssl()` param should override the presence
of an ssl config, meaning that we can apply this without being a breaking
change.