cleanup temp tables
Necessary for a corresponding internal change.
The general gist is open atlas never wanted this application-specific garbage anyways,
so I've mostly removed it here and made it so someone can just implement an interface.

Real reason for all of this is that we're moving to lazily initialized temp tables
because postgres' rubbish temp table creation speeds screw with Hikari pool resizing,
among other things.