Batch sweep cells (squashing commits from PR #1068)
Users can now configure both row batching (as before) and cell batching
(which is a very large value by default).

We may eventually want to eliminate one of these options (for example,
eliminate row batching in favor of cell batching); however, doing so is
not trivial because we currently have logic in place to dynamically
adjust the row batch size, so we leave this as future work.

Merge conflicts "solved" but there are likely to be compile errors in this commit...