Don't check nodetool status when starting Cassandra nodes.
In the MultiCassandraPerformanceEteTest, we start by bringing up a
Cassandra cluster. In one of the tests, we take down a Cassandra node,
and then bring it up again. In both cases, nodetool status is used to
check that the cluster has three nodes.

To run nodetool status in the docker containers we were using lxc-attach
when running on public CircleCI. Unfortunately, this command fails with
the error "lxc-attach: attach.c: lxc_attach: 710 failed to get the init
pid". After pairing with Vlad, we were unable to find a solution to
this.

This commit removes the checking of nodetool status entirely, instead
just relying on the thrift port being open as an indicator that the
cluster is ready. This should be perfectly sufficient. If it isn't,
we could call nodetool (or cqlsh -c 'select now() from system') on the
host machine, using the -h and -p options. However, we'd need to take
care that nodetool or cqlsh was available on the host machine.