Include the string 'EteTest' in the name of the new tests.
To prevent tests being run *not* as part of the test suite they belong
to, we have adopted the convention of excluding all tests with 'EteTest'
in the name from the tests run by gradle.

Prior to this commit, MultiCassandraDoubleNodeDownTest and
MultiCassandraSingleNodeDownTest were being run independently of
MultiCassandraTestSuite, and so didn't have a docker cluster.