Improve perf of getRowsColumnRange
It looks like UNION ALL is not particularly performant, so it's
probably not a good idea to UNION ALL a number of queries equal
to the number of requested rows. Since it's guaranteed that all
column range selections will be identical except for possibly
the first and last row, we can merge all the queries for those
intermediate rows into a single query.