Alter timeouts to reduce flakes.
There are two modifications in this commit: firstly, we add a timeout to
our nodetool commands. I've found that on some runs, 'nodetool status',
which is run to check that the Cassandra cluster has three nodes, can
hang indefinitely. I think this is due to an interplay between JMX and
Docker.

The second modification is that I am bumping the timeout on checking
that any Containers object is ready to two minutes. This is because the
ThreeNodeCassandraCluster object can legitimately take longer than a
minute to be ready, since the time includes a nodetool repair.