Move duplicated code into CassandraVersion
Every time we were getting exactVersion, we were immediately putting it
back into an environment variable. Now, we can call
CassandraVersion.getEnvironment() directly.

ThreeNodeCassandraCluster now contains both fromEnvironment() and
getEnvironment(). If the former is changed, e.g. to from("3.7"), it
could lead to inconsistencies. That can potentially be solved by
setting the environment variable in fromEnvironment(), if it did not
already exist.