Save timestamp-related thread dumps to temp files
Previously, these were outputted as regular logs, but this caused users
of Atlas-backed products to be needlessly worried, particularly as a
certain product triggers the warning during upgrades (and we believe
this to be benign).

Now, the thread dumps are outputted to unique files in a temporary.
During testing, these files were only 13KB, so it would take a lot of
triggers before the build-up of old thread dumps causes memory pressure.

A fairly easy alternative would be to save to a hard-coded temporary
file (and clear it if it was already present).