Add ability to use current AtlasDB version inside documentation
This commit uses ./gradlew printVersion to set the 'version' property of
Sphinx, so that |version| can be used to refer to the codebase's current
version throughout our documentation. Unfortunately, Sphinx doesn't
allow you to use this inside code-blocks, so I've added a custom
directive to pre-transform the content of code-blocks.