In the Jepsen framework, pass values as strings.
Previously we would pass the values returned by :ok events of our Jepsen
tests as longs. This was fine when our Jepsen tests were just for
timestamps, but since we want to extend the testing to locks, we need to
handle a wider variety of responses from the timelock server.

I've opted to use nullable strings, rather than Objects, since we really
don't need the extra complexity of fully-fledged objects.

[no release notes]