Add CheckAndSet to KeyValueService API
This is required for the backup lock to safely acquire and release
locks in a KVS-agnostic way.

So far, CheckAndSetRequest is designed to handle only a single cell at
a time, although it is built in such a way as to allow extension to
changing multiple cells at a time.

Like putUnlessExists, it seems unrealistic to support atomicity of CAS
across multiple cells, which should hopefully be made explicit if we
retain the model of only CAS-ing a single Cell-Value pair per request.

In this initial commit, non-trivial implementations have been stubbed,
and the associated tests (included in this commit) will fail. The next
series of commits should add support for each KVS in turn.