Prevent double-newlines when the last block contains no imports
Previously, if there were imports in some types, but not in the last
block, we would end up with a blank line at the end of the
`renderImports` part of the buffer.

This never occurred in practice, but a parallel PR, #1357, adds an
import type ".net", which doesn't usually have any matching imports.
Regenerating schemas lead to many undesirable blank lines.

We passed the responsibility of adding the line between imports and the
class declaration to ImportRenderer, because otherwise we'd need
omniscient logic along the lines of "only add a blank line if we just
rendered some imports *and* there will be more imports to come".

[no release notes]