Fix stream stores with large in-memory size
This fixes several issues with the current StreamStore implementation.

The implementation on the BlockGetter was inconsistent in how it used
expectedLength(). In some places it assumed it was the size of a single
block, and in others it assumed it was the maximum memory size for the
stream buffer. I have renamed this to expectedBlockLength and made the
usage consistent to be just the block size (and also match the javadoc).

The BlockConsumingInputStream was always allocating a buffer with a size
of the maximum number of blocks allowed in memory rather than the size
corresponding to the actual number of blocks that were going to be
loaded.

Added StreamTests that actually test the case where the
inMemoryThreshold is Integer.MAX_VALUE (as it is in some product code).
These tests fail without the corresponding fixes mentioned above.

[no release notes]