Don't preload buffer when creating a BlockConsumingInputStream
Fixes #1443.

The usage of this stream is in AbstractGenericStreamStore.makeStream.
Crucially, refillBuffer creates a transaction, and
BlockConsumingInputStream.create is itself called inside a transaction.

Having one transaction entirely inside another is not desirable because
of the potential for deadlock situations to occur, so this broke our
own best practices.

It will now be the caller's responsibility to close the "load stream"
transaction before reading the stream data.