Fix deadlocks in parallel DbKvs read operations
BatchedDbReadTable/LazyClosableIterator deadlock if the connection pool
is small enough: if the last available connection is held by the callable
at the tail of the queue, we won't be able to process the result of the
callable at the head (because that requires a connection). But the
connection can only be released once we get to the end of the queue,
which is impossible until we process the first element, hence the deadlock.

The solution is do parallelism at a higher level. If each thread
reads all results into memory and releases the connection before
returning, then at least one thread will make progress as long as at
least one connection is available.

We make DbReadTable a concrete class containing all the code from
UnbatchedDbReadTable and AbstractDbReadTable classes, and remove
BatchedDbReadTable. Instead we implement a simple "fork-join" framework
in ParallelTaskRunner.