Compressed streams: use a BlockGetter to load blocks
This commit introduces a 'brother' of
AbstractGenericStreamStore.makeStream to the compressed stream store
(see StreamTestWithHashStreamStore). This is done to avoid having to
keep the entire stream in memory at once.

As a result of this, a couple of imports are added to the generated
StreamStore files. Additionally, BlockGetter is now a public interface
- I slapped a javadoc to the header to explain when it should be used.