Bypass CheckAndSetException if persisted_locks was populated for us
There is a race condition in LockStore.create(), where two threads
could both check allLockEntries(), both see the set as empty, and then
both do the checkAndSet from (empty) to "lock open".

The loser of this race will see a CheckAndSetException, but all we
actually care about is that some thread managed to populate the table.
Therefore, we simply log.