Only report the reason when acquire fails
We want to make it slightly harder for users to release a lock that is
not theirs. This commit means they can still do this, but will have to
dive into the logs, rather than being able to do it programatically.

Rolling back a lock in this manner is not supposed to be a common
occurrence; it should only be needed if backup or sweep crashed. This
should be rare, and in any case would likely require support (or
log-diving by power users).