make sweep use deleteRange instead of truncate
Using deleteRange instead of truncate in sweep means that it no longer
conflicts with running online backups for dbkvs. As a workaround so that
performance of sweep is not terrible in Cassandra, added behavior so
that the Cassandra KVS turns deleteRange into a truncate when the range
represents the entire table. In the future the Cassandra KVS should
implement a better deleteRange for the case when truncate can't be used.