Declare TransactionManager.close will throw no checked exceptions
Currently, no implementation of TransactionManager.close throws any checked exceptions, but TransactionManager inherits AutoClose's close signature, which is declared as throwing Exception. This makes it hard to use, as catching all checked exceptions is not safe in a try-with-resources block, since it masks other checked exceptions declared in the block, and even when calling close in a dedicated try/catch block, users will have to handle InterruptedException separately (calling `Thread.currentThread().interrupt()` before wrapping the exception), leading to a lot of unnecessary code bloat.