cap transaction cache and add stats logging
This changes the cache to have a compound key so that individual cells
are cached so that a sane max size (in cells) can be enforced.

Also adds stats so we can determine if this cache is actually useful.

[no release notes]

change CachingTransaction logging to debug

reorder helper methods in CachingTransaction

fix CachingTransaction test and add tests

Clean up CachingTransactionTest

- Extract method to avoid duplicate code
- Extract constants for "row/col/value".getBytes()
- Rename variables to satisfy checkstyle
- Other checkstyle fixes