Single-query getRange() on DbKvs (#1805)
* Single-query getRange() on DbKvs

Currently, we do range scans in two steps on DbKvs: first, we load the
range of row_names, and then call getRows() to load the contents of
those rows. This is an extra round trip to the database that involves
sending lots of data back and forth. On top of that, we have to sort
the results in Java because they come back in random order, which burns
more CPU.

This is inefficient, and we can do the whole operation with one SQL query.

In addition, we now set the fetch size correctly (on the
PreparedStatement instead of the ResultSet). By the time we get a
ResultSet back, we already have the first page pre-fetched. Note that we
don't set the fetch size on Postgres because it has no effect:
see https://jdbc.postgresql.org/documentation/head/query.html

On Oracle, this results in the following changes in KvsGetRangeBenchmarks:
    getSingleRange[Dirty]: ~1.7x speedup
    getSingleLargeRange: ~4x speedup
    getSingleLargeRangeDirty: ~1.8x speedup

On Postgres, the new query is not executed in parallel anymore, so some
benchmarks will actually get a bit slower:
    getSingleRange[Dirty]: ~2x speedup
    getSingleLargeRange: ~10% slowdown
    getSingleLargeRangeDirty: ~20% slowdown

Even though the latency gets slightly worse for large ranges, the
throughput is increased significantly: the getSingleLargeRange[Dirty]
benchmarks load ranges of size 1000 which results in 4 batches being
executed in parallel (the default fetch batch size is 256 rows).
The new implementation achieves comparable latency without any parallelism.
I think this is the right trade-off to make.

* Make checkstyle happy and add a release note

* Add more tests for getRange() and fix the bugs they caught

Plus other minor changes

* Restructure PostgresGetRange.PageIterator.computeNext() a bit

* Use isFirstRowName in getNextStartRowUnlessTerminal

(cherry picked from commit 75e1cd2524a69308758fab8c670c0594f69cada6)