Refuse to start on Postgres 9.5.0 and 9.5.1 (#1820)
Refuse to start on Postgres 9.5.0 and 9.5.1

Postgres 9.5 introduced a bug that was fixed in 9.5.2 that causes
incorrect results to be returned for some queries. Here is a simple test
case:

create table foo(a int, b int, primary key (a, b));
insert into foo (a, b) values (1, 1);
select * from foo where (a, b) >= (1, 2); -- returns 0 rows (correct)
select * from foo where (a, b) >= (1, 2) and b = 1; -- returns 1 row (incorrect)

This breaks the new getRange query, so we should refuse to start
on the broken versions.