The big sweep rewrite, part 1
We want to replace KVS.getRangeOfTimestamps() with a new call
getCandidateCellsForSweeping(). In this commit, to keep the change
smaller, we temporarily implement the new call using two existing
calls (getRangeOfTimestamps() and getRange()) without changing the
KeyValueService interface. See GetCandidateCellsForSweepingShim.java.
This is more or less equivalent to the current implementation, just
written differently.

What's different about the new call:
    - It returns individual cells rather than RowResults, so it's more
      clear that large rows can be split into several batches;
    - Its behavior is very sweep-specific and allows only loading
      cells that can potentially be swept. This pushes more logic
      into the underlying database and allows sending less data back
      over the network;
    - The implementation for DbKvs will be much more efficient by using
      the same idea as the recent DbKvs.getRange() rewrite -
      performing a single query to get a single page of results;
    - If the THOROUGH strategy is used, the new call will return the
      flag that indicates whether the value corresponding to the latest
      timestamp is empty. Thus we can avoid calling getRange() to load
      values.

Also, we now re-partition the cells before we delete them. There are
now three parameters that define how batching is performed. See
SweepBatchConfig for the definition of these parameters. We also get rid
of the dynamic batch size multiplier in the background sweeper. It
didn't really make much sense: if you OOM, you crash hard anyway; if
some other error occurs, it's either a bug in the code or an
intermittent problem (such as network outage), and decreasing the batch
size would help neither.

Here is the plan for the next few commits:
    - Make the new call a part of the KVS interface, but still
      using the "shim" as the implementation;
    - Implement the call for DbKvs;
    - Implement the call for CassandraKvs;
    - Implement the call for InMemoryKvs;
    - (Optional) Implement the call for RocksDbKvs & JdbcKvs?
    - Delete getRangeOfTimestamps() because it shouldn't be used anywhere
      at that point.