fix getRange with column selection in console (#1876)
Console had a bug where a getRange with a column selection would fail
as soon as it tried to get the second batch. This was due to an
asymmetry in serialize/deserialize for TableRanges.

The TableRangeDeserializer converts long column names into short names,
and now (after this change) the TableRangeSerializer converts short
column names back into the long names.