correctly persist compression block size (#1995)
* correctly persist compression block size

TableMetadata previously did not ever persist the compression block
size, which meant that requests for explicit compression previously
had zero effect.

* add release notes for compression block size fix

* compression block sizes must be powers of 2

Cassandra rejects any attempts to set this to something other than a
power of 2, so add a check that will reject this to avoid cases where
a user does not realize their schema is invalid because it works on
non-Cassandra KVSes.

* more thoroughly test metadata serialization

Adds another test case that explicitly sets all the other fields (like
conflict handler, sweep strategy, etc) to make sure that they are
included in serialization.