Deprecate atlasdb-commons/AssertUtils defaultLog
Previously, failed AssertUtils asserts logged to a common logger.
However, this led to the problem that the logger would always be com.palantir.util.AssertUtils,
which is extremely annoying to try to filter based on your logging properties.
(Should it go into the server error log, or maybe the upgrade log, or the import log file?
Can't tell, cause it's all AssertUtils!)

This commit changes the API to specify an explicit logger.
Until we get all upstream projects off of using defaultLog however,
the existing methods will stay, just deprecated.