Fix deadlock encountered in internal security product
Attempting to connect to ourself to get a timestamp while processing a request can lead to deadlock if incoming request volume is high, as all Jetty threads end up waiting for the timestamp server, and no threads remain to actually handle the timestamp server requests. If there is a single leader, it must be us, and we can avoid the deadlock entirely; the leader block is only there to allow other services reading the configuration to find the timestamp server.

HA instances will have to switch to using the standalone timestamp server to fix this issue, as it is unavoidable there.