Fix deadlock encountered in internal security product
Attempting to connect to ourself to get a timestamp while processing a request can lead to deadlock if incoming request volume is high, as all Jetty threads end up waiting for the timestamp server, and no threads remain to actually handle the timestamp server requests. If we are the only leader, we can avoid the deadlock entirely by using the local server. Detect this situation by registeringa `timestamp/server-id` endpoint and querying it; if it returns the same UUID we generated locally,we are talking to ourself.

HA instances will have to switch to using the standalone timestamp server to fix this issue, asit is unavoidable there.