Fix deadlock encountered in internal security product (#2091)
* Remove unused field

* Add a Feign proxy factory that throws javax.ws.rs exceptions

Error handling with SerializableErrorDecoder is hard because it can throw arbitrary exception types, and falls back to RuntimeExceptions with the status code embedded in the error message string. Instead, clone some code from glassfish's Jersey client to convert Feign error responses to javax.ws.rs exceptions, which have a sensible type hierarchy for clean exception handling code.

* Fix deadlock encountered in internal security product

Attempting to connect to ourself to get a timestamp while processing a request can lead to deadlock if incoming request volume is high, as all Jetty threads end up waiting for the timestamp server, and no threads remain to actually handle the timestamp server requests. If we are the only leader, we can avoid the deadlock entirely by using the local server. Detect this situation by registeringa `timestamp/server-id` endpoint and querying it; if it returns the same UUID we generated locally,we are talking to ourself.

HA instances will have to switch to using the standalone timestamp server to fix this issue, asit is unavoidable there.

* Make async thread daemon threads

* Add release note

* Produces application/JSON

* Start, not run, the thread

* Reduce warning frequency to once a minute

* Move routing logic to createRawServices

* Peer review improvements

 * Use existing PingableLeader.getUUID instead of a new service
 * Use AwaitingLeadershipProxy even if there is only a single leader, to work around a bug in Atlas
 * Modify LocalOrRemoteProxy to retry on NotCurrentLeaderException

* Review nits

* Can't static import