Log query string in retry message
This is an attempt to restore the pertinent information lost by the
stack trace. If a deployment is seeing many timeouts that are
successfully retried, they may be able to determine the culprit by
examining the query strings, and concurrently checking the product's
slow-logs (timed out queries will very likely be logged there).