fix logging for cassandra case
The previous fix would fail for Cassandra multiputs because the
tableName string was not actually a real table name, but a set
of multiple table names.  For now changing this to just be an
unsafe arg and renaming the variable to hopefully avoid confusion
until this code can be refactored correctly.