Fix incorrect call to createFromFullyQualifiedName (#2407)
* Fix incorrect call to createFromFullyQualifiedName

Not all tables have namespaces. Whenever you have a tablename as a
string you cannot assume that it represents a table that has a
namespace.

This fixes an IllegalArgumentException that would get thrown whenever
this particular logging was triggered on a table that had no namespace.

[no release notes]

* fix logging for cassandra case

The previous fix would fail for Cassandra multiputs because the
tableName string was not actually a real table name, but a set
of multiple table names.  For now changing this to just be an
unsafe arg and renaming the variable to hopefully avoid confusion
until this code can be refactored correctly.

* Remove unused imports