Recover when our view of the world gets outdated
It can happen that releasing the persistent lock fails, but the
database actually completes the CAS operation (after timing out waiting
for an "ack").

Previously when this happened, we would never be able to acquire the
lock again, because we'd maintain a belief that we hold lockId.

In this commit, we attempt to acquire the lock *even if we think we
hold it anyway*. There are three possibilities, covered by individual
tests.

Case 1 (callingAcquireTwiceFails): we still hold the lock. we fail with
IllegalStateException, as the caller should call release... before
calling acquire again.

Case 2 (canAcquireAfterReleaseSeemsToFailButSecretlySucceeds): the lock
is available, so we should be able to grab it.

Case 3
(cannotAcquireAfterReleaseFailsButSecretlySucceedsAndThenSomeoneElseTake
sTheLock): the lock was taken, but by someone else. We keep retrying,
but set the lockId to null (as we definitely don't hold it).