Prevent sweep getting stuck when PersistentLock release fails (#2419)
* Handle persistent locks being cleared from under us

Fixes an issue where, if the persistent lock is cleared from under us,
we would still hold the view that we have it, and would therefore never
acquire another lock.

* Clarify log message on release failure

Database errors would throw some error other than CheckAndSetException.

* Add PR numbers

* Make log message less scary.

Hallowe'en isn't until the end of the month, after all...

* Recover when our view of the world gets outdated

It can happen that releasing the persistent lock fails, but the
database actually completes the CAS operation (after timing out waiting
for an "ack").

Previously when this happened, we would never be able to acquire the
lock again, because we'd maintain a belief that we hold lockId.

In this commit, we attempt to acquire the lock *even if we think we
hold it anyway*. There are three possibilities, covered by individual
tests.

Case 1 (callingAcquireTwiceFails): we still hold the lock. we fail with
IllegalStateException, as the caller should call release... before
calling acquire again.

Case 2 (canAcquireAfterReleaseSeemsToFailButSecretlySucceeds): the lock
is available, so we should be able to grab it.

Case 3
(cannotAcquireAfterReleaseFailsButSecretlySucceedsAndThenSomeoneElseTake
sTheLock): the lock was taken, but by someone else. We keep retrying,
but set the lockId to null (as we definitely don't hold it).

* SafeArgs

* KvsBackedPersistentLockService: use WebPreconditions

Throws a BadRequestException instead of an NPE.

* Fix test name

* Move release note to correct place

* Return success when we try to acquire having held the lock all along

Practically speaking, the following must have happened:
1. We acquired the lock successfully
2. We failed to talk to the DB when releasing the lock
3. We recovered ability to talk to the DB, and it still has our lock
from 1

We want and have the lock - sounds good! Return success, but log that
this happened. Since we can talk to the DB, we would expect the
subsequent release call to succeed.

* LockStoreImpl: add SafeArgs

* Fix NPE and add debug logging

* Fix Errorprone warnings