Wire up STM's initializer with its last prerequisite
Two caveats here as I made the smallest possible change:

1. TransactionManagerInitializer is not visible from
atlasdb-impl-shared, where the SerializableTransactionManager lives. We
could move it up the dependency tree, but I like it where it is, and
the STM only needs the initializer portion of it as a prerequisite.

2. Indeed, only the InitializeCheckingWrapper part of STM needs the
initializer part of TMI, so I added a prerequisite object to its
wrapper.