Maintain order when grouping cells into rows
Collectors.groupingBy uses a HashMap by default, which does not
preserve the order of elements.

Using a LinkedHashMap retains the ordering (which we already rely on
earlier in the pipeline, right after fetching stuff from CQL).
If we need to add ordering here, one way would be to use a TreeMap, but
this would hurt performance - as the list is already sorted, insertion
would become O(log n) rather than O(1) as it is now.