Cassandra getCandidateCellsForSweeping (#2546)
* revert cassandra sweep changes (#2521)

* revert cassandra sweep changes

* add back some tests

* move tests to cassandra

* fix batchsize

* newline

* delete tests

* release notes

* Nziebart/cassandra sweep batch size (#2525)

* cassandra default sweep candidate batch size

* docs

* checkpoint

* checkpoint

* finish up

* restore tests

* fix build

* checkstyle

* unused classes

* fix getRangeOfTimestamps

* push row grouping into getCandidateCellsForSweeping

* Revert "push row grouping into getCandidateCellsForSweeping"

This reverts commit 898b7763e96b0aa8abeca38facb30bc4ffc647db.

* rename

* batch sizes

* javadocs

* fix build

* fix build

* extra tests

* PR comments

* minor stuff

* Fix compile

* Maintain order when grouping cells into rows

Collectors.groupingBy uses a HashMap by default, which does not
preserve the order of elements.

Using a LinkedHashMap retains the ordering (which we already rely on
earlier in the pipeline, right after fetching stuff from CQL).
If we need to add ordering here, one way would be to use a TreeMap, but
this would hurt performance - as the list is already sorted, insertion
would become O(log n) rather than O(1) as it is now.

* valuesBatchSize must be > 0

* fix nits

* Fix test setup

* Actually take a peek at the PeekingIterator

* Catch UnavailableException and put it on ICE

We say we throw InsufficientConsistencyException on the KVS API, so
we'd best stick to it.

* checkstyle