Rename LockRefreshingTimelockService to TimeLockClient
As a deleted comment says, this is the single class where we should
talk to TimeLock.

However, this class does have two responsibilities:
1 - send requests to TimeLock (handling connection exceptions)
2 - refresh locks

We considered the trade-offs of pulling out the 'refresh locks' part
into an AutoDelegate, at the cost of adding another layer of wrapping
and delegation, and decided for now to keep both in the same class. We
have enough wrapping layers that the confusion of having both things in
one small class is probably less than that of having another layer.