Don't loop infinitely if there's no data left in the table
Fix an edge case that caused an infinite loop in
AbstractSweepTaskRunnerTest.testSweepLatestDeletedThrough.

If the range contains a row that has no columns (including if the row's
only column was deleted), the Thrift rowGetter query will return that
row, although the list of columns would be empty - the row's still
there, but it's tombstoned. Therefore, determineSafeRangeEndInclusive
would return that row's key as a safe range end (and not
Optional.Empty). So we run CqlExecutor.getTimestamps, get an empty
list, and iterate again through the loop.

Previously, we set rangeStart to the previous rangeEnd. Both ends of
the CqlExecutor.getTimestamps query are inclusive, so on the next
iteration we'd get the empty row again, and rangeEnd and rangeStart
would never change.

Using RangeRequests.nextLexicographicName avoids fetching the same
empty row again, so we can continue making forward progress.