Remove "ttl" from Cell and delete ExpiringKVS (#2599)
* Remove "ttl" from Cell and delete ExpiringKVS

This has never been used by anyone. If we actually wanted to support the
expiration dates in Cassandra, the Cell class would be a wrong place
to put the TTL because Cell represents a key, and TTL is not part of
a cell key. The fact that TTL is not part of the equals() and hashCode()
in that class is quite telling.

Not sure why Cell is marked as Serializable. I don't think any of our
products sends objects of this class over the network using Java
serialization or persists them in a database.

* Add release notes for #2599

* Fix SnapshotTransaction.put()

* Remove expiration from TableMetadata and code gen

- Regenerate TableMetadataPersistence protobufs
- Regenerate all schemas

* Fix TracingKeyValueServiceTest that relies on Cell.toString()

* Fix compile after merge

* Make Parameter final Again

* Pad out and rebase release note