Store and reuse host for prepared query
Since the query executor is created by CqlExecutorImpl, which itself is
recreated anew for each "get candidate cells for sweeping" iteration
(and only used in sweep), we expect this map to contain exactly one
element.

If for some reason the map is not filled, we make a best guess at the
right host for this query.